@@ -23,12 +23,15 @@ from __future__ import annotations
 
 from types import FunctionType
 from types import WrapperDescriptorType
+from typing import TYPE_CHECKING
 from typing import Any
-from typing import Callable
 
 from docstring_inheritance.docstring_inheritors.bases.inheritor import (
     BaseDocstringInheritor,
 )
+
+if TYPE_CHECKING:
+    from collections.abc import Callable
 
 DocstringInheritorClass = type[BaseDocstringInheritor]
 
